+++ R/nonlinear.R
@@ -417,7 +417,7 @@ S7::method(term_build, NlTerm) <- function(term, data, ...) {
           name = paste0(p, "::", e$name),
           index = index[[p]][e$index],
           penalty = e$penalty, fixed = e$fixed, n_values = e$n_values,
-          min_ratio = e$min_ratio)
+          values = e$values, min_ratio = e$min_ratio)
       }
     }
   }
@@ -501,6 +501,7 @@ S7::method(term_build, NlTerm) <- function(term, data, ...) {
       pens[[length(pens) + 1L]] <- list(name = key, index = off + e$index,
                                         penalty = e$penalty, fixed = e$fixed,
                                         n_values = e$n_values,
+                                        values = e$values,
                                         min_ratio = e$min_ratio)
     }
     off <- off + ncol(mats[[lb]])